ui_builder/requirements.txt
fastapi>=0.115
uvicorn>=0.34
httpx>=0.28
python-multipart>=0.0.18
pyyaml>=6.0.1
